feat: 진료비 계산 구현
