--- src/main/java/com/example/hospitalreservation/common/message/Message.java
@@ -1,6 +1,0 @@
-package com.example.hospitalreservation.common.message;
-
-public interface Message {
-
-    String getMessage();
-}
