--- src/main/java/com/example/hospitalreservation/common/treatment/DefaultFeeCalculator.java
@@ -1,12 +1,0 @@
-package com.example.hospitalreservation.common.treatment;
-
-public class DefaultFeeCalculator {
-
-    public static final FeeCalculator SUM = (TreatmentPurpose... purposes) -> {
-        int total = 0;
-        for (TreatmentPurpose purpose : purposes) {
-            total += purpose.getFee();
-        }
-        return total;
-    };
-}
