--- src/main/java/com/example/hospitalreservation/common/treatment/FeeCalculator.java
@@ -1,7 +1,0 @@
-package com.example.hospitalreservation.common.treatment;
-
-@FunctionalInterface
-public interface FeeCalculator {
-
-    int calculate(TreatmentPurpose... purposes);
-}
