--- src/main/java/com/example/hospitalreservation/common/treatment/Purpose.java
@@ -1,8 +1,0 @@
-package com.example.hospitalreservation.common.treatment;
-
-public interface Purpose {
-
-    String getType();
-
-    int getFee();
-}
